refactor parser utils
