+++ scripts/src/plh-data-convert/utils/plh-string.utils.ts
